+++ CHANGELOG.md
@@ -6,6 +6,13 @@ Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how 
 
 ## [Unreleased]
 
+## [0.1.3] - 2024-12-11
+
+### Fixed
+- use `ssh2.utils.generateKeyPair()` instead of local ssh client (closes #10)
+- Use external browser for help (closes #3)
+- WAGO W Logo changed to a square 256x256 image (closes #8)
+
 ## [0.1.2] - 2024-11-29
 
 ### Added